WIP refactor app component hooks
